+++ tests/InvalidConfigTest.php
@@ -22,7 +22,7 @@ class InvalidConfigTest extends \PHPUnit_Framework_TestCase
         $previous = new \RuntimeException('msg');
         $e = new InvalidConfig('MyConfig', 'oh, error', 2, $previous);
         $this->assertSame('MyConfig', $e->getConfigName());
-        $this->assertSame('oh, error', $e->getErrmsg());
+        $this->assertSame('oh, error', $e->getErrorMessage());
         $this->assertSame(2, $e->getCode());
         $this->assertSame($previous, $e->getPrevious());
         $this->assertSame('MyConfig has an invalid format: "oh, error"', $e->getMessage());